+++ dart/lib/src/platform/platform.dart
@@ -1,42 +1,30 @@
-import 'dart:typed_data';
+import '_io_platform.dart' if (dart.library.js_interop) '_web_platform.dart'
+    as impl;
 
-import '_io_platform.dart'
-    if (dart.library.html) '_html_platform.dart'
-    if (dart.library.js_interop) '_web_platform.dart' as platform;
-
-const Platform instance = platform.instance;
-
-abstract class Platform {
+class Platform extends impl.PlatformBase {
   const Platform();
 
-  /// A string (`linux`, `macos`, `windows`, `android`, `ios`, or `fuchsia`)
-  /// representing the operating system.
-  String get operatingSystem;
+  bool get isLinux => operatingSystem == OperatingSystem.linux;
 
-  /// A string representing the version of the operating system or platform.
-  String get operatingSystemVersion;
+  bool get isMacOS => operatingSystem == OperatingSystem.macos;
 
-  /// Get the local hostname for the system.
-  String get localHostname;
+  bool get isWindows => operatingSystem == OperatingSystem.windows;
 
-  /// Endianness of this platform.
-  Endian get endian => Endian.host;
+  bool get isAndroid => operatingSystem == OperatingSystem.android;
 
-  /// True if the operating system is Linux.
-  bool get isLinux => (operatingSystem == 'linux');
+  bool get isIOS => operatingSystem == OperatingSystem.ios;
 
-  /// True if the operating system is OS X.
-  bool get isMacOS => (operatingSystem == 'macos');
+  bool get isFuchsia => operatingSystem == OperatingSystem.fuchsia;
 
-  /// True if the operating system is Windows.
-  bool get isWindows => (operatingSystem == 'windows');
+  bool get supportsNativeIntegration => !isFuchsia;
+}
 
-  /// True if the operating system is Android.
-  bool get isAndroid => (operatingSystem == 'android');
-
-  /// True if the operating system is iOS.
-  bool get isIOS => (operatingSystem == 'ios');
-
-  /// True if the operating system is Fuchsia
-  bool get isFuchsia => (operatingSystem == 'fuchsia');
+enum OperatingSystem {
+  android,
+  fuchsia,
+  ios,
+  linux,
+  macos,
+  windows,
+  unknown,
 }