+++ dart/test/sentry_test.dart
@@ -5,10 +5,9 @@ import 'package:sentry/src/dart_exception_type_identifier.dart';
 import 'package:sentry/src/event_processor/deduplication_event_processor.dart';
 import 'package:test/test.dart';
 
-import 'fake_platform_checker.dart';
 import 'mocks.dart';
 import 'mocks/mock_integration.dart';
-import 'mocks/mock_platform_checker.dart';
+import 'mocks/mock_runtime_checker.dart';
 import 'mocks/mock_sentry_client.dart';
 import 'test_utils.dart';
 
@@ -136,6 +135,14 @@ void main() {
     });
 
     test('should start transaction with context', () async {
+      final tr = Sentry.startTransactionWithContext(
+          SentryTransactionContext('name', 'operation'));
+      await tr.finish();
+
+      expect(client.captureTransactionCalls.length, 1);
+    });
+
+    test('should start transaction with hint', () async {
       final tr = Sentry.startTransactionWithContext(
           SentryTransactionContext('name', 'operation'));
       await tr.finish();
@@ -364,7 +371,7 @@ void main() {
 
     test('should set options.debug to true when in debug mode', () async {
       final options = defaultTestOptions();
-      options.platformChecker = MockPlatformChecker(isDebug: true);
+      options.runtimeChecker = MockRuntimeChecker(isDebug: true);
 
       expect(options.debug, isFalse);
       await Sentry.init(
@@ -378,7 +385,7 @@ void main() {
 
     test('should respect user options.debug when in debug mode', () async {
       final options = defaultTestOptions();
-      options.platformChecker = MockPlatformChecker(isDebug: true);
+      options.runtimeChecker = MockRuntimeChecker(isDebug: true);
 
       expect(options.debug, isFalse);
       await Sentry.init(
@@ -394,7 +401,7 @@ void main() {
     test('should leave options.debug unchanged when not in debug mode',
         () async {
       final options = defaultTestOptions();
-      options.platformChecker = MockPlatformChecker(isDebug: false);
+      options.runtimeChecker = MockRuntimeChecker(isDebug: false);
 
       expect(options.debug, isFalse);
       await Sentry.init(
@@ -435,7 +442,8 @@ void main() {
   });
 
   test('options.environment debug', () async {
-    final sentryOptions = defaultTestOptions(FakePlatformChecker.debugMode());
+    final sentryOptions =
+        defaultTestOptions(checker: MockRuntimeChecker(isDebug: true));
     await Sentry.init(
       (options) {
         options.dsn = fakeDsn;
@@ -447,7 +455,8 @@ void main() {
   });
 
   test('options.environment profile', () async {
-    final sentryOptions = defaultTestOptions(FakePlatformChecker.profileMode());
+    final sentryOptions =
+        defaultTestOptions(checker: MockRuntimeChecker(isProfile: true));
 
     await Sentry.init(
       (options) {
@@ -460,7 +469,8 @@ void main() {
   });
 
   test('options.environment production (defaultEnvironment)', () async {
-    final sentryOptions = defaultTestOptions(FakePlatformChecker.releaseMode());
+    final sentryOptions =
+        defaultTestOptions(checker: MockRuntimeChecker(isRelease: true));
     await Sentry.init(
       (options) {
         options.dsn = fakeDsn;
@@ -472,24 +482,22 @@ void main() {
   });
 
   test('options.logger is set by setting the debug flag', () async {
-    final sentryOptions = defaultTestOptions(FakePlatformChecker.debugMode());
+    final sentryOptions =
+        defaultTestOptions(checker: MockRuntimeChecker(isDebug: true));
 
     await Sentry.init(
       (options) {
         options.dsn = fakeDsn;
         options.debug = true;
-        // ignore: deprecated_member_use_from_same_package
         expect(options.logger, isNot(noOpLogger));
 
         options.debug = false;
-        // ignore: deprecated_member_use_from_same_package
         expect(options.logger, noOpLogger);
       },
       options: sentryOptions,
     );
 
-    // ignore: deprecated_member_use_from_same_package
-    expect(sentryOptions.logger, isNot(dartLogger));
+    expect(sentryOptions.logger, noOpLogger);
   });
 
   group('Sentry init optionsConfiguration', () {
@@ -500,9 +508,9 @@ void main() {
     });
 
     test('throw is handled and logged', () async {
-      // Use release mode in platform checker to avoid additional log
+      // Use release mode in runtime checker to avoid additional log
       final sentryOptions =
-          defaultTestOptions(FakePlatformChecker.releaseMode())
+          defaultTestOptions(checker: MockRuntimeChecker(isRelease: true))
             ..automatedTestMode = false
             ..debug = true
             ..logger = fixture.mockLogger;