+++ flutter/lib/src/sentry_flutter_options.dart
@@ -1,7 +1,5 @@
 import 'dart:async';
 
-import 'package:file/file.dart';
-import 'package:file/local.dart';
 import 'package:flutter/services.dart';
 import 'package:flutter/widgets.dart';
 import 'package:meta/meta.dart' as meta;
@@ -12,7 +10,6 @@ import 'event_processor/screenshot_event_processor.dart';
 import 'navigation/time_to_display_tracker.dart';
 import 'renderer/renderer.dart';
 import 'screenshot/sentry_screenshot_quality.dart';
-import 'sentry_flutter.dart';
 import 'sentry_privacy_options.dart';
 import 'sentry_replay_options.dart';
 import 'user_interaction/sentry_user_interaction_widget.dart';
@@ -21,7 +18,7 @@ import 'user_interaction/sentry_user_interaction_widget.dart';
 /// Note that some of these options require native Sentry integration, which is
 /// not available on all platforms.
 class SentryFlutterOptions extends SentryOptions {
-  SentryFlutterOptions({super.dsn, super.checker}) {
+  SentryFlutterOptions({super.dsn, super.platform, super.checker}) {
     enableBreadcrumbTrackingForCurrentPlatform();
   }
 
@@ -175,14 +172,6 @@ class SentryFlutterOptions extends SentryOptions {
   /// Enable auto performance tracking by default.
   bool enableAutoPerformanceTracing = true;
 
-  /// Automatically track app start measurement and send it with the
-  /// first transaction. Set to false when configuring option to disable or if
-  /// you want to set the end time of app startup manually using
-  /// [SentryFlutter.setAppStartEnd].
-  @Deprecated(
-      'Will be removed in v9. In order to disable app starts disable it via option.removeIntegration(...) instead')
-  bool autoAppStart = true;
-
   /// Automatically attaches a screenshot when capturing an error or exception.
   ///
   /// Requires adding the [SentryWidget] to the widget tree.
@@ -193,14 +182,6 @@ class SentryFlutterOptions extends SentryOptions {
 
   /// The quality of the attached screenshot
   SentryScreenshotQuality screenshotQuality = SentryScreenshotQuality.high;
-
-  /// Only attach a screenshot when the app is resumed.
-  /// See https://docs.sentry.io/platforms/flutter/troubleshooting/#screenshot-integration-background-crash
-  bool attachScreenshotOnlyWhenResumed = false;
-
-  @Deprecated(
-      'Will be removed in a future version. Use [beforeCaptureScreenshot] instead')
-  BeforeScreenshotCallback? beforeScreenshot;
 
   /// Sets a callback which is executed before capturing screenshots. Only
   /// relevant if `attachScreenshot` is set to true. When false is returned
@@ -303,12 +284,12 @@ class SentryFlutterOptions extends SentryOptions {
   /// you must use `SentryWidgetsFlutterBinding.ensureInitialized()` instead.
   bool enableFramesTracking = true;
 
-  /// Controls initialization of the Sentry Javascript SDK on web platforms.
-  /// When enabled and [autoInitializeNativeSdk] is true, loads and initializes
-  /// the JS SDK in the document head.
-  ///
-  /// Defaults to `false`
-  bool enableSentryJs = false;
+  /// Replay recording configuration.
+  final replay = SentryReplayOptions();
+
+  /// Privacy configuration for masking sensitive data in screenshots and Session Replay.
+  /// Screen content masking is enabled by default.
+  final privacy = SentryPrivacyOptions();
 
   /// By using this, you are disabling native [Breadcrumb] tracking and instead
   /// you are just tracking [Breadcrumb]s which result from events available
@@ -350,7 +331,7 @@ class SentryFlutterOptions extends SentryOptions {
   /// available in the Flutter environment. This way you get more detailed
   /// information where available.
   void enableBreadcrumbTrackingForCurrentPlatform() {
-    if (platformChecker.hasNativeIntegration) {
+    if (platform.supportsNativeIntegration) {
       useNativeBreadcrumbTracking();
     } else {
       useFlutterBreadcrumbTracking();
@@ -394,55 +375,7 @@ class SentryFlutterOptions extends SentryOptions {
   @override
   // ignore: invalid_use_of_internal_member
   set automatedTestMode(bool value) => super.automatedTestMode = value;
-
-  @meta.internal
-  FileSystem fileSystem = LocalFileSystem();
-
-  /// Configuration of experimental features that may change or be removed
-  /// without prior notice. Additionally, these features may not be ready for
-  /// production use yet.
-  @meta.experimental
-  final experimental = _SentryFlutterExperimentalOptions();
 }
-
-class _SentryFlutterExperimentalOptions {
-  /// Replay recording configuration.
-  final replay = SentryReplayOptions();
-
-  /// Privacy configuration for masking sensitive data in screenshots and Session Replay.
-  /// Screen content masking is:
-  /// - enabled by default for SessionReplay
-  /// - disabled by default for screenshots captured with events.
-  /// In order to mask screenshots captured with events, access or change
-  /// this property in your application: `options.experimental.privacy`.
-  /// Doing so will indicate that you want to configure privacy settings and
-  /// will enable screenshot masking alongside the default replay masking.
-  /// Note: this will change in a future SDK major release to enable screenshot
-  /// masking by default for all captures.
-  SentryPrivacyOptions get privacy {
-    // If the user explicitly sets the privacy setting, we use that.
-    // Otherwise, we use the default settings, which is no masking for screenshots
-    // and full masking for session replay.
-    // This property must only by accessed by user code otherwise it defeats the purpose.
-    _privacy ??= SentryPrivacyOptions();
-    return _privacy!;
-  }
-
-  /// TODO: remove when default masking value are synced with SS & SR in the next major release
-  SentryPrivacyOptions? _privacy;
-
-  @meta.internal
-  SentryPrivacyOptions? get privacyForScreenshots => _privacy;
-
-  @meta.internal
-  SentryPrivacyOptions get privacyForReplay =>
-      _privacy ?? SentryPrivacyOptions();
-}
-
-@Deprecated(
-    'Will be removed in a future version. Use [BeforeCaptureCallback] instead')
-typedef BeforeScreenshotCallback = FutureOr<bool> Function(SentryEvent event,
-    {Hint? hint});
 
 /// A callback which can be used to suppress capturing of screenshots.
 /// It's called in [ScreenshotEventProcessor] if screenshots are enabled.