+++ flutter/test/event_processor/url_filter/web_url_filter_event_processor_test.dart
@@ -1,5 +1,5 @@
 @TestOn('browser')
-library flutter_test;
+library;
 
 import 'package:flutter_test/flutter_test.dart';
 import 'package:sentry_flutter/sentry_flutter.dart';